refactor(core): extract asset_download module from scraper_service (#443)
Move download_assets_if_enabled verbatim into application::asset_download.
It stays in the application layer: it reads ScraperConfig, extracts asset
URLs and delegates the transfer to the AssetDownloaderPort adapter. Pushing
it into adapters::downloader would invert the dependency direction, so it is
kept as orchestration and re-exported from scraper_service so discovery.rs
and the integration tests resolve unchanged. Completes the #443 decomposition
— scraper_service now holds only scrape_with_config, scrape_multiple_with_limit
and ScrapeOutcome. Pure refactoring.
